src/main/webapp/app/entities/rss/route/rss-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IRss, Rss } from '../rss.model';
import { RssService } from '../service/rss.service';

import { RssRoutingResolveService } from './rss-routing-resolve.service';

describe('Rss routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RssRoutingResolveService;
  let service: RssService;
  let resultRss: IRss | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(RssRoutingResolveService);
    service = TestBed.inject(RssService);
    resultRss = undefined;
  });

  describe('resolve', () => {
    it('should return IRss returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRss = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRss).toEqual({ id: 123 });
    });

    it('should return new IRss if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRss = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultRss).toEqual(new Rss());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Rss })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRss = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultRss).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
